Python Projects/Flask/Daily Code App/requirements.txt
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
Click==7.0
Django==2.2.4
django-crispy-forms==1.7.2
dominate==2.3.5
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-Ext==0.1
Flask-Gunicorn==0.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gcloud==0.18.3
googleapis-common-protos==1.6.0
gunicorn==19.9.0
httplib2==0.13.0
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
oauth2client==4.1.3
Pillow==6.1.0
protobuf==3.9.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pymongo==3.8.0
python-docx==0.8.10
python-dotenv==0.10.3
pytz==2019.2
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.5
sqlparse==0.3.0
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
